requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nose==1.3.7
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
SQLAlchemy==1.4.44
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
